requirements.txt
# ==============================
# Core Inference
# ==============================

huggingface_hub>=0.23.4
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
numpy==1.26.4

# ==============================
# Utilities
# ==============================

python-dotenv==1.0.1
tqdm==4.66.4
packaging==24.0
typing_extensions==4.10.0

# ==============================
# Networking
# ==============================

httpx==0.27.0
httpcore==1.0.5
anyio==4.3.0
certifi==2024.2.2
idna==3.6

# ==============================
# SentenceTransformer dependency
# ==============================

transformers==4.40.2
tokenizers==0.19.1
safetensors==0.4.2